* Created simple wix installer that builds MSI for the EDI Project using the contents of the Release folder
Customised the install dialog workflow so License dialog is switched off.
Modified the installer project file - added before and after build steps, that harvest the contents of the release folder and moves the MSI file to a Install directory. Added banners for the MSI dialogs
Added links.txt which contain useful urls that helped me develop this solution.

mbullock976